requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
attrs==20.3.0
Authlib==0.15.3
autopep8==1.5
Babel==2.8.0
Beaker==1.11.0
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==8.0.1
cryptography==3.2.1
defusedxml==0.6.0
Flask==2.0.0
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Cors==3.0.8
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.1
flask-swagger-ui==3.25.0
flask-talisman==0.8.1
Flask-WTF==1.0.0
future==0.18.2
google-auth==2.3.3
gunicorn==20.0.4
idna==2.9
importlib-metadata==3.1.0
inflect==4.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
lasio==0.25.1
Mako==1.1.4
MarkupSafe==2.0.0
marshmallow==3.12.1
ntlm-auth==1.5.0
numpy==1.19.0
oic==1.2.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pycryptodomex==3.9.9
pyjwkest==1.4.2
pymongo==3.11.2
PyMySQL==1.0.2
pyop==3.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2020.1
requests==2.23.0
requests-ntlm==1.1.0
secure-smtplib==0.1.1
six==1.14.0
speaklater==1.3
sqlacodegen==2.1.0
SQLAlchemy==1.3.15
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.9
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.0
simplekv==0.14.1